feat: run markdownlint for PRs (#120)
Use Github actions to run markdownlint for PRs. A reasonable set of
defaults are chosen and configured under `.markdownlint.json`.

Keeping the `.markdownlint.json` file in the root of the folder so that
it is picked up by VS Code markdownlint plugin by default.

Motivations:
The main motivations is so that all RFCs meet a reasonable bar in terms
of consistency. The main violator is line width - unusually long lines
are hard to review and comment on in PRs. They also don't play very well
with text editors.

The added benefit is that some parts of markdown are flimsy such that
they need to be formatted in a particular way to be rendered correctly.
This will help users keep to that.