RFC 249: Rewrite the V2 Experiments RFC to have a more working-backwards tone (#377)
While working on the release of the CDK v2 Experimental Modules, I found myself looking for a place to document and discuss some of the customer-facing decisions being made (e.g., package naming, versioning, Changelog organization). Unfortunately, the current state of RFC 249 does not lend itself well to this format (IMO); the RFC as-is is more implementation-focused and more heavily leans towards the why of many of the decisions. While absolutely useful context, it doesn't read as much as a working-backwards document as I'd like.

My mental model to solve this was to copy (and save!) this really useful decision-making document, and replace the main RFC body with a series of working-backwards artifacts that (hopefully?) capture the customer experience of working with the alpha/unstable modules.