chore: update CDKv2 RFC guidance for v1 dist tags post-launch (#386)
The RFC currently specifies that post-launch of CDK v2, the v2 packages will get
an NPM distribution tag of `latest`, and the v1 packages will get a tag of
`v1`. However, tags that can be interpreted as valid semver ranges are not
allowed, and it is impossible to publish without a tag. Rename the tag to
`latest-1` to get around this. This is the same convention `npm` uses for its
releases.