Add customer pains to the RFC
IMHO it will be much easier to read the RFC if it will include a short description of the customer pains at the beginning of the RFC, before getting into details and without mentioning any implementation details. What is the pain we are trying to solve here? Why are we even having this RFC?