chore: explain linting process
Move the markdownlint config file to the top of the repository, so that
the VSCode extension can find the rules while you are editing. Explain
to install the extension (or run the command) in the top-level README.

Fix the linting problems on the main README that there are
spaces inside a link (MD039).

Rewrite the `lint.sh` script to be simpler: we don't need to use `find`,
and in this way it can accurately reflect the finding of issues in the
exit code.