RFC 732: CDK CLI Telemetry (#734)
This is a request for comments about CDK CLI Telemetry. See #732  for
additional details. 

**The purpose of this RFC is to iterate in public. We welcome community
feedback on the opt-out mechanism,
on whether you would decide to opt-out, or on additional telemetry
values that we are missing that have proven helpful in other CLIs.**

APIs are signed off by @iliapolo 

---

_By submitting this pull request, I confirm that my contribution is made
under
the terms of the Apache-2.0 license_

---------

Co-authored-by: Momo Kornher <kornherm@amazon.co.uk>
Co-authored-by: Otavio Macedo <288203+otaviomacedo@users.noreply.github.com>
Co-authored-by: Eli Polonsky <epolon@amazon.com>