RFC 762: Serverless ElastiCache L2 Construct Improvement (#803)
This is an improvement (update some examples to be similar to the implementation code) of the request for comments about introducing an L2 Construct for `Serverless ElastiCache` together with the other needed constructs: `User` and `UserGroup`. See #762 for additional details. 

APIs are signed off by @otaviomacedo.

---

_By submitting this pull request, I confirm that my contribution is made under
the terms of the Apache-2.0 license_